+++ Projects/App/Sources/Application/NeedleGenerated.swift
@@ -325,8 +325,17 @@ private class StorageDependency1447167c38e97ef97427Provider: StorageDependency {
     var signInFactory: any SignInFactory {
         return appComponent.signInFactory
     }
-    var afterLoginComponent: AfterLoginComponent {
-        return appComponent.afterLoginComponent
+    var myPlayListComponent: MyPlayListComponent {
+        return appComponent.myPlayListComponent
+    }
+    var multiPurposePopUpFactory: any MultiPurposePopUpFactory {
+        return appComponent.multiPurposePopUpFactory
+    }
+    var favoriteComponent: FavoriteComponent {
+        return appComponent.favoriteComponent
+    }
+    var textPopUpFactory: any TextPopUpFactory {
+        return appComponent.textPopUpFactory
     }
     private let appComponent: AppComponent
     init(appComponent: AppComponent) {
@@ -399,40 +408,6 @@ private class MyPlayListDependency067bbf42b28f80e413acProvider: MyPlayListDepend
 /// ^->AppComponent->MyPlayListComponent
 private func factory51a57a92f76af93a9ec2f47b58f8f304c97af4d5(_ component: NeedleFoundation.Scope) -> AnyObject {
     return MyPlayListDependency067bbf42b28f80e413acProvider(appComponent: parent1(component) as! AppComponent)
-}
-private class AfterLoginDependencya880b76858e0a77ed700Provider: AfterLoginDependency {
-    var fetchUserInfoUseCase: any FetchUserInfoUseCase {
-        return appComponent.fetchUserInfoUseCase
-    }
-    var logoutUseCase: any LogoutUseCase {
-        return appComponent.logoutUseCase
-    }
-    var requestComponent: RequestComponent {
-        return appComponent.requestComponent
-    }
-    var profilePopComponent: ProfilePopComponent {
-        return appComponent.profilePopComponent
-    }
-    var myPlayListComponent: MyPlayListComponent {
-        return appComponent.myPlayListComponent
-    }
-    var multiPurposePopUpFactory: any MultiPurposePopUpFactory {
-        return appComponent.multiPurposePopUpFactory
-    }
-    var favoriteComponent: FavoriteComponent {
-        return appComponent.favoriteComponent
-    }
-    var textPopUpFactory: any TextPopUpFactory {
-        return appComponent.textPopUpFactory
-    }
-    private let appComponent: AppComponent
-    init(appComponent: AppComponent) {
-        self.appComponent = appComponent
-    }
-}
-/// ^->AppComponent->AfterLoginComponent
-private func factory6cc9c8141e04494113b8f47b58f8f304c97af4d5(_ component: NeedleFoundation.Scope) -> AnyObject {
-    return AfterLoginDependencya880b76858e0a77ed700Provider(appComponent: parent1(component) as! AppComponent)
 }
 private class FavoriteDependency8f7fd37aeb6f0e5d0e30Provider: FavoriteDependency {
     var containSongsFactory: any ContainSongsFactory {
@@ -799,7 +774,6 @@ extension AppComponent: Registration {
         localTable["fetchNewSongsUseCase-any FetchNewSongsUseCase"] = { [unowned self] in self.fetchNewSongsUseCase as Any }
         localTable["signInFactory-any SignInFactory"] = { [unowned self] in self.signInFactory as Any }
         localTable["storageComponent-StorageComponent"] = { [unowned self] in self.storageComponent as Any }
-        localTable["afterLoginComponent-AfterLoginComponent"] = { [unowned self] in self.afterLoginComponent as Any }
         localTable["requestComponent-RequestComponent"] = { [unowned self] in self.requestComponent as Any }
         localTable["localAuthDataSource-any LocalAuthDataSource"] = { [unowned self] in self.localAuthDataSource as Any }
         localTable["remoteAuthDataSource-any RemoteAuthDataSource"] = { [unowned self] in self.remoteAuthDataSource as Any }
@@ -986,7 +960,10 @@ extension ServiceInfoComponent: Registration {
 extension StorageComponent: Registration {
     public func registerItems() {
         keyPathToName[\StorageDependency.signInFactory] = "signInFactory-any SignInFactory"
-        keyPathToName[\StorageDependency.afterLoginComponent] = "afterLoginComponent-AfterLoginComponent"
+        keyPathToName[\StorageDependency.myPlayListComponent] = "myPlayListComponent-MyPlayListComponent"
+        keyPathToName[\StorageDependency.multiPurposePopUpFactory] = "multiPurposePopUpFactory-any MultiPurposePopUpFactory"
+        keyPathToName[\StorageDependency.favoriteComponent] = "favoriteComponent-FavoriteComponent"
+        keyPathToName[\StorageDependency.textPopUpFactory] = "textPopUpFactory-any TextPopUpFactory"
     }
 }
 extension FaqComponent: Registration {
@@ -1010,18 +987,6 @@ extension MyPlayListComponent: Registration {
         keyPathToName[\MyPlayListDependency.deletePlayListUseCase] = "deletePlayListUseCase-any DeletePlayListUseCase"
         keyPathToName[\MyPlayListDependency.logoutUseCase] = "logoutUseCase-any LogoutUseCase"
         keyPathToName[\MyPlayListDependency.textPopUpFactory] = "textPopUpFactory-any TextPopUpFactory"
-    }
-}
-extension AfterLoginComponent: Registration {
-    public func registerItems() {
-        keyPathToName[\AfterLoginDependency.fetchUserInfoUseCase] = "fetchUserInfoUseCase-any FetchUserInfoUseCase"
-        keyPathToName[\AfterLoginDependency.logoutUseCase] = "logoutUseCase-any LogoutUseCase"
-        keyPathToName[\AfterLoginDependency.requestComponent] = "requestComponent-RequestComponent"
-        keyPathToName[\AfterLoginDependency.profilePopComponent] = "profilePopComponent-ProfilePopComponent"
-        keyPathToName[\AfterLoginDependency.myPlayListComponent] = "myPlayListComponent-MyPlayListComponent"
-        keyPathToName[\AfterLoginDependency.multiPurposePopUpFactory] = "multiPurposePopUpFactory-any MultiPurposePopUpFactory"
-        keyPathToName[\AfterLoginDependency.favoriteComponent] = "favoriteComponent-FavoriteComponent"
-        keyPathToName[\AfterLoginDependency.textPopUpFactory] = "textPopUpFactory-any TextPopUpFactory"
     }
 }
 extension FavoriteComponent: Registration {
@@ -1205,7 +1170,6 @@ private func registerProviderFactory(_ componentPath: String, _ factory: @escapi
     registerProviderFactory("^->AppComponent->FaqComponent", factory4e13cc6545633ffc2ed5f47b58f8f304c97af4d5)
     registerProviderFactory("^->AppComponent->QuestionComponent", factoryedad1813a36115eec11ef47b58f8f304c97af4d5)
     registerProviderFactory("^->AppComponent->MyPlayListComponent", factory51a57a92f76af93a9ec2f47b58f8f304c97af4d5)
-    registerProviderFactory("^->AppComponent->AfterLoginComponent", factory6cc9c8141e04494113b8f47b58f8f304c97af4d5)
     registerProviderFactory("^->AppComponent->FavoriteComponent", factory8e4acb90bd0d9b48604af47b58f8f304c97af4d5)
     registerProviderFactory("^->AppComponent->RequestComponent", factory13954fb3ec537bab80bcf47b58f8f304c97af4d5)
     registerProviderFactory("^->AppComponent->NoticeDetailComponent", factory3db143c2f80d621d5a7fe3b0c44298fc1c149afb)
