+++ Projects/Features/StorageFeature/Sources/Reactors/StorageReactor.swift
@@ -1,0 +1,34 @@
+import Foundation
+import ReactorKit
+
+final class StorageReactor: Reactor {
+    
+    enum Action {
+        
+    }
+    
+    enum Mutation {
+        
+    }
+    
+    struct State {
+        
+    }
+    
+    let initialState: State
+    
+    init() {
+        initialState = State (
+        
+        )
+    }
+    
+    func mutate(action: Action) -> Observable<Mutation> {
+        
+    }
+    
+    func reduce(state: State, mutation: Mutation) -> State {
+        
+    }
+    
+}
